backend/requirements.txt
Flask
python-dotenv
requests
pymongo
cryptography
Flask-CORS
